Fix: Handle null dereference issue in directory iterator usage (2)
:Release Notes:
Handle null dereference issue in directory iterator usage (2)

:Detailed Notes:
- Fix Static Analysis issue (CID 3269600)
- Depending on the platform, the directory iterator may return
  a null pointer when the directory is empty. This patch adds
  a check to handle the null pointer case.

:Testing Performed:
Locally built.

:QA Notes:
N/A

:Issues Addressed:
N/A

Reviewed-on: http://gpro.lge.com/c/webosose/edge-ai-computer-vision/+/408508
Tested-by: Junil Kim <jjunil79.kim@lge.com>

+++ core/base/ExtensionLoader.cpp
@@ -56,6 +56,8 @@ t_aif_status ExtensionLoader::init(bool readRegistryFile, std::string pluginPath
   try {
     for (const auto &entry : fs::directory_iterator(pluginPath))
     {
+      if (!entry)
+        continue;
       Logi("ExtensionLoader::init()", "pluginPath:", pluginPath, ", entry.path:", entry.path().c_str());
       if (!entry.is_regular_file() ||
           fs::is_symlink(entry.path()) ||
@@ -431,6 +433,8 @@ bool ExtensionLoader::isNeededToGenRegistryFile()
       try {
         for (const auto &entry : fs::directory_iterator(dir))
         {
+          if (!entry)
+            continue;
           if (entry.path().string().find(m_registryFilePath+".done.") == 0)
           {
             fs::remove(entry.path());
